рефакторинг, создал дирректорию для функций и кнопок общего назначения common_functions
